Refactor: Tests, TransactionService method create and TransactionItemService.

+++ src/main/java/com/gustavosass/finance/service/TransactionService.java
@@ -6,8 +6,10 @@ import java.util.NoSuchElementException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.gustavosass.finance.exceptions.FoundItemsPaidForTransactionException;
 import com.gustavosass.finance.model.Account;
 import com.gustavosass.finance.model.Transaction;
+import com.gustavosass.finance.model.TransactionItem;
 import com.gustavosass.finance.repository.TransactionRepository;
 
 @Service
@@ -30,24 +32,28 @@ public class TransactionService {
 		return transactionRepository.findById(id).orElseThrow(() -> new NoSuchElementException("Transaction not found."));
 	}
 	
-	public Transaction create(Transaction transaction) {
+	public Transaction create(Transaction transaction){
+
 		Account account = accountService.findById(transaction.getAccount().getId());
-
 		transaction.setAccount(account);
-		Transaction transactionCreated = transactionRepository.save(transaction);
-
-		transactionItemService.createTransactionsItems(transactionCreated);
-
-		return transactionCreated;
+		
+		return transactionRepository.save(transaction);
 	}
 
 	public Transaction update(Long id, Transaction transaction) {
 		Transaction transactionExists = transactionRepository.findById(id).orElseThrow(() -> new NoSuchElementException("Transaction not found."));
 		Account accountExists = accountService.findById(transaction.getAccount().getId());
+		
+		if (transactionExists.getInstallmentNumbers() != transaction.getInstallmentNumbers() || transactionExists.getValue() != transaction.getValue()) {
+			
+			List<TransactionItem> installmentsPaid = transactionItemService.listInstallmentsPaidByIdTransaction(transaction.getId());
 
-		transactionItemService.existsItemsPaidForTransaction(transaction.getId());
-		transactionItemService.findAllById(id).stream().forEach((item) -> transactionItemService.delete(id, item.getId()));
-		transactionItemService.createTransactionsItems(transaction);
+			if (installmentsPaid.size() > 0) {
+				throw new FoundItemsPaidForTransactionException("Founded installments paid, please cancel installments for exclude.");
+			}
+			
+			transactionItemService.findAllById(id).stream().forEach((item) -> transactionItemService.delete(id, item.getId()));
+		}		
 		
 		transactionExists.setAccount(accountExists);
 		transactionExists.setInstallmentNumbers(transaction.getInstallmentNumbers());
@@ -59,7 +65,7 @@ public class TransactionService {
 	
 	public void delete(Long id) {
 		transactionRepository.findById(id).orElseThrow(() -> new NoSuchElementException("Transaction not found."));
-		transactionItemService.existsItemsPaidForTransaction(id);
+		transactionItemService.listInstallmentsPaidByIdTransaction(id);
 		transactionRepository.deleteById(id);
 	}
 }
